fix(watch): stop command delivery after device revocation
Revalidate exact HTTP and node session ownership synchronously after pairing checks. Use the same boundary for polling, disconnect, and result commit, with real HTTP regression coverage and an unchanged extracted server fixture.
